+++ src/main/java/com/nextcloudlab/kickytime/match/controller/MatchingAdminController.java
@@ -1,10 +1,10 @@
-package com.nextcloudlab.kickytime.controller;
+package com.nextcloudlab.kickytime.match.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import com.nextcloudlab.kickytime.service.MatchingService;
+import com.nextcloudlab.kickytime.match.service.MatchingService;
 
 import lombok.RequiredArgsConstructor;
