+++ src/main/java/com/nextcloudlab/kickytime/match/dto/MatchCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.nextcloudlab.kickytime.match.controller;
+package com.nextcloudlab.kickytime.match.dto;
 
 import java.time.LocalDateTime;
