+++ src/main/java/com/nextcloudlab/kickytime/match/service/MatchParticipantService.java
@@ -1,13 +1,15 @@
 package com.nextcloudlab.kickytime.match.service;
 
-import com.nextcloudlab.kickytime.match.dto.MyMatchesResponse;
-import com.nextcloudlab.kickytime.match.entity.MatchStatus;
-import com.nextcloudlab.kickytime.match.repository.MatchParticipantRepository;
-import lombok.RequiredArgsConstructor;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+import com.nextcloudlab.kickytime.match.dto.MyMatchesResponse;
+import com.nextcloudlab.kickytime.match.entity.MatchStatus;
+import com.nextcloudlab.kickytime.match.repository.MatchParticipantRepository;
+
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
@@ -32,24 +34,19 @@ public class MatchParticipantService {
     }
 
     private long calculateUpcomingCount(List<MyMatchesResponse.MatchInfo> matches) {
-        return matches.stream()
-                .filter(this::isUpcoming)
-                .count();
+        return matches.stream().filter(this::isUpcoming).count();
     }
 
     private long calculateCompletedCount(List<MyMatchesResponse.MatchInfo> matches) {
-        return matches.stream()
-                .filter(this::isCompleted)
-                .count();
+        return matches.stream().filter(this::isCompleted).count();
     }
 
     private boolean isUpcoming(MyMatchesResponse.MatchInfo match) {
-        return  match.matchStatus() == MatchStatus.FULL ||
-                match.matchStatus() == MatchStatus.OPEN;
+        return match.matchStatus() == MatchStatus.FULL || match.matchStatus() == MatchStatus.OPEN;
     }
 
     private boolean isCompleted(MyMatchesResponse.MatchInfo match) {
-        return match.matchStatus() == MatchStatus.CLOSED ||
-               match.matchStatus() == MatchStatus.CANCELLED;
+        return match.matchStatus() == MatchStatus.CLOSED
+                || match.matchStatus() == MatchStatus.CANCELLED;
     }
 }
