+++ src/main/java/com/nextcloudlab/kickytime/match/service/MatchService.java
@@ -7,8 +7,8 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.nextcloudlab.kickytime.match.controller.MatchCreateRequestDto;
-import com.nextcloudlab.kickytime.match.controller.MatchResponseDto;
+import com.nextcloudlab.kickytime.match.dto.MatchCreateRequestDto;
+import com.nextcloudlab.kickytime.match.dto.MatchResponseDto;
 import com.nextcloudlab.kickytime.match.entity.Match;
 import com.nextcloudlab.kickytime.match.entity.MatchParticipant;
 import com.nextcloudlab.kickytime.match.entity.MatchStatus;
