+++ src/main/java/com/nextcloudlab/kickytime/match/service/MatchingService.java
@@ -1,4 +1,4 @@
-package com.nextcloudlab.kickytime.service;
+package com.nextcloudlab.kickytime.match.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
